% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CI1}
\alias{CI1}
\title{Calculate confidence interval}
\usage{
CI1(x)
}
\arguments{
\item{x}{Numeric vector for which to compute the confidence interval}
}
\value{
The computed confidence interval as a vector containing lower and upper bounds
}
\description{
Calculate confidence interval
}
\examples{
x <- c(1,2,3,4,4,5)
CI1(x)
}
