% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CI2}
\alias{CI2}
\title{Calculate S3 class 95\% confidence interval, lower and upper bounds}
\usage{
CI2(x)
}
\arguments{
\item{x}{Random numeric vector for which to compute the confidence interval}
}
\value{
Lower and upper bounds of the 95\% confidence intervals
}
\description{
Calculate S3 class 95\% confidence interval, lower and upper bounds
}
\examples{
set.seed(1234)
x <- rnorm(100)
obj <- make_ci_class(x)
print(obj)
CI2(obj)
}
