% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fn_sin}
\alias{fn_sin}
\title{Compute sine using truncated series expansion}
\usage{
fn_sin(x, k)
}
\arguments{
\item{x}{The number for sine computation}

\item{k}{Number of terms beyond the x term in the series expansion}
}
\value{
The computed sine value
}
\description{
Compute sine using truncated series expansion
}
\examples{
fn_sin(x=0.5, k=5)
}
