% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_ci_class}
\alias{make_ci_class}
\title{Constructor function for ci_class}
\usage{
make_ci_class(x)
}
\arguments{
\item{x}{Random numeric vector for which to compute the confidence interval}
}
\value{
The computed constructor function
}
\description{
Constructor function for ci_class
}
\examples{
set.seed(1234)
x <- rnorm(100)
obj <- make_ci_class(x)
}
