% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print}
\alias{print}
\title{Print ci_class}
\usage{
print(x)
}
\arguments{
\item{x}{Random numeric vector for which to compute the confidence interval}
}
\value{
Name of the class and number of observations in the S3 object
}
\description{
Print ci_class
}
\examples{
#' set.seed(1234)
x <- rnorm(100)
obj <- make_ci_class(x)
print(obj)
}
