% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_mean}
\alias{sample_mean}
\title{Calculate sample mean}
\usage{
sample_mean(...)
}
\arguments{
\item{x}{Numeric vector for which to compute the mean}
}
\value{
The computed sample mean
}
\description{
Calculate sample mean
}
\examples{
x <- c(1,2,3,4,4,5)
sample_mean(x)
}
