% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_sd}
\alias{sample_sd}
\title{Calculate sample standard deviation}
\usage{
sample_sd(x)
}
\arguments{
\item{x}{Numeric vector for which to compute the standard deviation}
}
\value{
The computed sample standard deviation
}
\description{
Calculate sample standard deviation
}
\examples{
x <- c(1,2,3,4,4,5)
sample_sd(x)
}
